requirements.txt
# Production dependencies
streamlit>=1.32.0
google-cloud-texttospeech>=2.15.0
python-dotenv>=1.0.0

# Development dependencies
flake8>=6.1.0
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.7.0
isort>=5.12.0
mypy>=1.5.1
